# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movie App API",
        "description": "Tài liệu API cho hệ thống đặt vé xem phim",
        "contact": {
            "email": "youremail@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/create-payment": {
            "post": {
                "tags": [
                    "Thanh toán MoMo"
                ],
                "summary": "Tạo yêu cầu thanh toán MoMo",
                "operationId": "2472906c5b05f1e508ef231bf8b53897",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "orderId"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100000
                                    },
                                    "orderId": {
                                        "type": "string",
                                        "example": "VE123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Thành công, trả về thông tin thanh toán",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Payment processed successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi trong việc tạo yêu cầu thanh toán",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Không thể tạo yêu cầu thanh toán"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/callback": {
            "post": {
                "tags": [
                    "Thanh toán MoMo"
                ],
                "summary": "Callback MoMo khi thanh toán thành công hoặc thất bại",
                "operationId": "ea517113726a80afbbded65a03fd10b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "resultCode",
                                    "orderId"
                                ],
                                "properties": {
                                    "resultCode": {
                                        "type": "string",
                                        "example": "0"
                                    },
                                    "orderId": {
                                        "type": "string",
                                        "example": "VE123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Thanh toán thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Payment processed successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Thanh toán thất bại",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/checkTrasaction": {
            "post": {
                "tags": [
                    "Thanh toán MoMo"
                ],
                "summary": "Thông báo IPN từ MoMo về trạng thái thanh toán",
                "operationId": "7dd1fc5a1703261cbe00955d9c509f0e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "orderId"
                                ],
                                "properties": {
                                    "orderId": {
                                        "type": "string",
                                        "example": "VE123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Thanh toán đã được xác nhận",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Payment processed successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Thanh toán không thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "fail"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Payment failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Không thể kiểm tra trạng thái giao dịch",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unable to check transaction status"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}